+++ gnat/clients/__init__.py
@@ -39,6 +39,7 @@ from gnat.connectors.opencti.client import OpenCTIClient
 from gnat.connectors.qradar.connector import QRadarConnector
 from gnat.connectors.sentinel.connector import SentinelConnector
 from gnat.connectors.wazuh.connector import WazuhConnector
+from gnat.connectors.servicenow.client import ServiceNowClient
 
 CLIENT_REGISTRY: dict = {
     "threatq":        ThreatQClient,
@@ -71,6 +72,7 @@ CLIENT_REGISTRY: dict = {
     "qradar":         QRadarConnector,
     "sentinel":       SentinelConnector,
     "wazuh":          WazuhConnector,
+    "servicenow":     ServiceNowClient,
 }
 
 __all__ = [